Move IngestJobStatusStore into ingest-core
